third_party/WebKit/LayoutTests/platform/win/external/wpt/navigation-timing/test_unique_performance_objects-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
PASS window.performance.timing is defined 
PASS Different window objects have unique performance objects 
Harness: the test ran to completion.
